Abstracet out interface rendering to new class InterfaceRenderer
