+++ src/test/java/biblioteket/roborally/elements/interactingelements/FlagTest.java
@@ -1,4 +1,4 @@
-package biblioteket.roborally.elements.interactingElements;
+package biblioteket.roborally.elements.interactingelements;
 
 import biblioteket.roborally.actors.IPlayer;
 import biblioteket.roborally.actors.IRobot;
@@ -7,14 +7,13 @@ import biblioteket.roborally.actors.Robot;
 import biblioteket.roborally.board.DirVector;
 import biblioteket.roborally.board.Direction;
 import biblioteket.roborally.elements.ArchiveMarkerElement;
-import biblioteket.roborally.elements.interactingelements.SingleWrenchRepairElement;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class WrenchRepairTest {
+public class FlagTest {
     private IPlayer player;
 
     @BeforeEach
@@ -24,24 +23,37 @@ public class WrenchRepairTest {
     }
 
     @Test
-    void singleWrenchRepairElementRemovesOneDamageTokenTest() {
-        SingleWrenchRepairElement repair = new SingleWrenchRepairElement();
-        // Add 5 damage tokens to robot
-        player.getRobot().addDamageTokens(5);
+    void interactingWithCorrectFlagRegistersFlagTest() {
+        FlagElement flag = new FlagElement(1);
+        flag.interact(player);
 
-        repair.interact(player);
-
-        assertEquals(4, player.getRobot().getNumberOfDamageTokens());
+        assertEquals(1, player.getNumberOfVisitedFlags());
     }
 
     @Test
-    void wrenchElementUpdatesArchiveMarkerTest() {
-        SingleWrenchRepairElement repair = new SingleWrenchRepairElement();
+    void cantPickupFlag2BeforeFlag1Test() {
+        FlagElement flag1 = new FlagElement(1);
+        FlagElement flag2 = new FlagElement(2);
+
+        flag2.interact(player);
+        assertEquals(0, player.getNumberOfVisitedFlags());
+
+        flag1.interact(player);
+        assertEquals(1, player.getNumberOfVisitedFlags());
+
+        flag2.interact(player);
+        assertEquals(2, player.getNumberOfVisitedFlags());
+
+    }
+
+    @Test
+    void touchingFlagUpdatesArchiveMarkerTest() {
+        FlagElement flag = new FlagElement(1);
         IRobot robot = player.getRobot();
         DirVector newPosition = new DirVector(5, 5, Direction.NORTH);
 
         robot.setPosition(newPosition);
-        repair.interact(player);
+        flag.interact(player);
 
         assertEquals(newPosition.getX(), robot.getArchiveMarker().getX());
         assertEquals(newPosition.getY(), robot.getArchiveMarker().getY());
