+++ src/test/java/biblioteket/roborally/elements/interactingelements/HoleTest.java
@@ -1,11 +1,10 @@
-package biblioteket.roborally.elements.interactingElements;
+package biblioteket.roborally.elements.interactingelements;
 
 import biblioteket.roborally.actors.IPlayer;
 import biblioteket.roborally.actors.IRobot;
 import biblioteket.roborally.actors.Player;
 import biblioteket.roborally.actors.Robot;
 import biblioteket.roborally.elements.ArchiveMarkerElement;
-import biblioteket.roborally.elements.interactingelements.HoleElement;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
